requirements.txt
# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1

# Excel file processing
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# OpenAI API
openai==1.54.0
httpx==0.27.0

# HTTP requests for external APIs
requests==2.31.0

# SQL Server database connection
pyodbc==5.0.1

# Environment variable management
python-dotenv==1.0.0

# Azure SDK (for deployment)
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0

# Date handling
python-dateutil==2.8.2

# Data validation
marshmallow==3.20.1

# Word document generation
python-docx==1.1.0

# PDF processing
PyPDF2==3.0.1

# SharePoint integration
Office365-REST-Python-Client==2.5.3

# Production WSGI server
gunicorn==21.2.0
